Snafu removed in order to limit the dependencies
The project was completely refactored to make it easier to maintain.
Also Snafu crate was removed in order to limit the number of
dependencies of the project.

Snafu is really awesome but I think I don't need it for the moment. It's
better to handle error using From trait implementations manually.
