+++ Cargo.toml
@@ -12,7 +12,6 @@ keywords = ["api", "github", "tool"]
 [dependencies]
 serde_json = "1.0.79"
 serde_yaml = "0.8.23"
-snafu = "0.7.0"
 hyper-tls = "0.5.0"
 rpassword = "6.0.1"
 rprompt = "1.0.5"
